Swap shape_function.py to element_style.py
