+++ python/src/main.py
@@ -1,9 +1,12 @@
 import matplotlib.pyplot as plt
 import numpy as np
+import time
 
 import io_data
 import input_wave
 import plot_model
+
+start = time.time()
 
 ## --- Input FEM Mesh --- ##
 fem = io_data.input_mesh("input/mesh.in")
@@ -16,7 +19,7 @@ fem.set_output(outputs)
 
 ## --- Define input wave --- ##
 fsamp = 5000
-duration = 1.5
+duration = 0.02
 
 tim,dt = np.linspace(0,duration,int(fsamp*duration),endpoint=False,retstep=True)
 wave_acc = input_wave.tapered_sin(tim,fp=5.0,taper=0.2,duration=1.0,amp=1.0)
@@ -25,7 +28,7 @@ ntim = len(tim)
 
 ## --- Prepare time solver --- ##
 fem.update_init(dt)
-ax = plot_model.plot_mesh_update_init()
+# ax = plot_model.plot_mesh_update_init()
 
 ## Iteration ##
 output_vel = np.zeros((ntim,fem.output_nnode))
@@ -41,11 +44,15 @@ for it in range(len(tim)):
     output_strain[it,:] = [element.strain[0] for element in fem.output_elements]
 
     if it%10 == 0:
-        plot_model.plot_mesh_update(ax,fem,500.)
+        # plot_model.plot_mesh_update(ax,fem,500.)
         print(it,output_vel[it,0],output_strain[it,0])
 
+
+elapsed_time = time.time() - start
+print ("elapsed_time: {0}".format(elapsed_time) + "[sec]")
+
 ## Output result ##
-plt.figure()
-plt.plot(tim,wave_vel)
-plt.plot(tim,output_vel[:,0])
-plt.show()
+# plt.figure()
+# plt.plot(tim,wave_vel)
+# plt.plot(tim,output_vel[:,0])
+# plt.show()
